requirements.txt
aiogram>=3.4.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
pillow>=10.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
qrcode>=8.0
sqlalchemy>=2.0.0
